+++ lychee-lib/src/types/base_info.rs
